Added Application
Added the VSCode project with the application.

This is a JavaFX project that uses fxml to separate the frontend and the backend, namely by having the frontend coded up by fxml and the backend programmed in java in separate files.

This version has the primary menu, with all the DB manipulation buttons just loading up an empty secondary window.

D2L template java file for connecting to oracle DB is in the project but project is missing the other .jar file needed.

Errors in this version all come from the oracle connection template java file; the actual project is fine.